# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_hanasakis_kotoha_pkg_ollama.Message:
    properties:
      content:
        type: string
      role:
        type: string
      tool_calls:
        items:
          $ref: '#/definitions/github_com_hanasakis_kotoha_pkg_ollama.ToolCall'
        type: array
    type: object
  github_com_hanasakis_kotoha_pkg_ollama.ToolCall:
    properties:
      function:
        $ref: '#/definitions/github_com_hanasakis_kotoha_pkg_ollama.ToolCallFunction'
    type: object
  github_com_hanasakis_kotoha_pkg_ollama.ToolCallFunction:
    properties:
      arguments:
        additionalProperties: true
        type: object
      name:
        type: string
    type: object
...
  internal_observability.ChatRequest:
    properties:
      history:
        items:
          $ref: '#/definitions/github_com_hanasakis_kotoha_pkg_ollama.Message'
        type: array
      message:
        type: string
    required:
    - message
    type: object
  internal_observability.EvalRun:
    properties:
      accuracy:
        type: number
      category:
        type: string
      created_at:
        type: string
      failed:
        type: integer
      id:
        type: integer
      metrics:
        description: JSON blob for detailed metrics
        type: string
      passed:
        type: integer
      total_tests:
        type: integer
    type: object
  internal_order.CreateOrderInput:
    properties:
      address_id:
        type: integer
      note:
        type: string
    required:
    - address_id
    type: object
  internal_order.Order:
    properties:
      created_at:
        type: string
      currency:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/internal_order.OrderItem'
        type: array
      order_no:
        type: string
      paid_at:
        type: string
      status:
        type: string
      stripe_session_id:
        type: string
      total_amount:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  internal_order.OrderItem:
    properties:
      id:
        type: integer
      name:
        type: string
      order_id:
        type: integer
      price:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      sku_id:
        type: integer
    type: object
  internal_payment.CheckoutInput:
    properties:
      cancel_url:
        type: string
      idempotency_key:
        type: string
      success_url:
        type: string
    required:
    - cancel_url
    - success_url
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: AI-powered snack e-commerce platform with hybrid search, conversational
    agent, and Stripe payments.
  title: Kotoha API
  version: "1.0"
paths:
...
  /admin/search/index:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Re-index all products into Milvus (admin)
      tags:
      - admin
  /agent/chat:
    post:
      consumes:
      - application/json
      parameters:
      - description: Chat message and history
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_observability.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Chat with AI shopping assistant
      tags:
      - agent
...
  /catalog/search:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 10
        description: Max results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 'Search products (hybrid: dense + BM25)'
      tags:
      - catalog
  /eval/runs:
    get:
      parameters:
      - description: Filter by category
        in: query
        name: category
        type: string
      - default: 20
        description: Max results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List evaluation runs
      tags:
      - observability
    post:
      consumes:
      - application/json
      parameters:
      - description: Evaluation run data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_observability.EvalRun'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_observability.EvalRun'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Save an evaluation run
      tags:
      - observability
...
  /orders:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List user's orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      parameters:
      - description: Order details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_order.CreateOrderInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_order.Order'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create an order from cart
      tags:
      - orders
  /orders/{id}:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_order.Order'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - orders
  /orders/{id}/cancel:
    post:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cancel an order
      tags:
      - orders
  /orders/{id}/checkout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Checkout options
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_payment.CheckoutInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create Stripe checkout session
      tags:
      - orders
...
  /webhook:
    post:
      consumes:
      - application/json
      parameters:
      - description: Stripe webhook signature
        in: header
        name: Stripe-Signature
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Handle Stripe webhook events
      tags:
      - payment
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"